# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Electronics Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: brandId
          in: query
          schema:
            type: integer
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: featured
          in: query
          schema:
            type: boolean
        - name: bestSeller
          in: query
          schema:
            type: boolean
        - name: newArrival
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, popularity]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product (admin)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/slug/{slug}:
    get:
      operationId: getProductBySlug
      tags: [products]
      summary: Get product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List testimonials
      responses:
        "200":
          description: Testimonials list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Create testimonial (admin)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestimonialBody"
      responses:
        "201":
          description: Testimonial created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /testimonials/{id}:
    put:
      operationId: updateTestimonial
      tags: [testimonials]
      summary: Update testimonial (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestimonialBody"
      responses:
        "200":
          description: Updated testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: deleteTestimonial
      tags: [testimonials]
      summary: Delete testimonial (admin)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: admin_session
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    ProductImage:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        imageUrl:
          type: string
        isMain:
          type: boolean
      required:
        - id
        - productId
        - imageUrl
        - isMain
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        shortDescription:
          type: string
          nullable: true
        fullDescription:
          type: string
          nullable: true
        price:
          type: string
        oldPrice:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true
        stockQuantity:
          type: integer
        stockStatus:
          type: string
          enum: [in_stock, low_stock, out_of_stock]
        featured:
          type: boolean
        bestSeller:
          type: boolean
        newArrival:
          type: boolean
        isPublished:
          type: boolean
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
        specifications:
          type: string
          nullable: true
        brandId:
          type: integer
          nullable: true
        categoryId:
          type: integer
          nullable: true
        brand:
          $ref: "#/components/schemas/Brand"
          nullable: true
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - slug
        - price
        - stockQuantity
        - stockStatus
        - featured
        - bestSeller
        - newArrival
        - isPublished
        - images
        - createdAt
        - updatedAt
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - limit
        - totalPages
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        shortDescription:
          type: string
          nullable: true
        fullDescription:
          type: string
          nullable: true
        price:
          type: string
        oldPrice:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true
        stockQuantity:
          type: integer
        stockStatus:
          type: string
          enum: [in_stock, low_stock, out_of_stock]
        featured:
          type: boolean
        bestSeller:
          type: boolean
        newArrival:
          type: boolean
        isPublished:
          type: boolean
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
        specifications:
          type: string
          nullable: true
        brandId:
          type: integer
          nullable: true
        categoryId:
          type: integer
          nullable: true
        images:
          type: array
          items:
            type: object
            properties:
              imageUrl:
                type: string
              isMain:
                type: boolean
            required:
              - imageUrl
              - isMain
      required:
        - name
        - price
        - stockQuantity
        - stockStatus
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        shortDescription:
          type: string
          nullable: true
        fullDescription:
          type: string
          nullable: true
        price:
          type: string
        oldPrice:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true
        stockQuantity:
          type: integer
        stockStatus:
          type: string
          enum: [in_stock, low_stock, out_of_stock]
        featured:
          type: boolean
        bestSeller:
          type: boolean
        newArrival:
          type: boolean
        isPublished:
          type: boolean
        seoTitle:
          type: string
          nullable: true
        seoDescription:
          type: string
          nullable: true
        specifications:
          type: string
          nullable: true
        brandId:
          type: integer
          nullable: true
        categoryId:
          type: integer
          nullable: true
        images:
          type: array
          items:
            type: object
            properties:
              imageUrl:
                type: string
              isMain:
                type: boolean
            required:
              - imageUrl
              - isMain
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        productCount:
          type: integer
      required:
        - id
        - name
        - slug
        - productCount
...
    Brand:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        logo:
          type: string
          nullable: true
        productCount:
          type: integer
      required:
        - id
        - name
        - slug
        - productCount
...
    Testimonial:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        rating:
          type: integer
        comment:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - name
        - rating
        - comment
        - isActive
    CreateTestimonialBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        rating:
          type: integer
        comment:
          type: string
        isActive:
          type: boolean
      required:
        - name
        - rating
        - comment
...