# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Electronics Store API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProductBodyImagesItem } from "./createProductBodyImagesItem";
import type { CreateProductBodyStockStatus } from "./createProductBodyStockStatus";

export interface CreateProductBody {
  name: string;
  slug?: string;
  shortDescription?: string | null;
  fullDescription?: string | null;
  price: string;
  oldPrice?: string | null;
  sku?: string | null;
  stockQuantity: number;
  stockStatus: CreateProductBodyStockStatus;
  featured?: boolean;
  bestSeller?: boolean;
  newArrival?: boolean;
  isPublished?: boolean;
  seoTitle?: string | null;
  seoDescription?: string | null;
  specifications?: string | null;
  brandId?: number | null;
  categoryId?: number | null;
  images?: CreateProductBodyImagesItem[];
}