# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Electronics Store API
 * OpenAPI spec version: 0.1.0
 */
import type { Brand } from "./brand";
import type { Category } from "./category";
import type { ProductImage } from "./productImage";
import type { ProductStockStatus } from "./productStockStatus";

export interface Product {
  id: number;
  name: string;
  slug: string;
  shortDescription?: string | null;
  fullDescription?: string | null;
  price: string;
  oldPrice?: string | null;
  sku?: string | null;
  stockQuantity: number;
  stockStatus: ProductStockStatus;
  featured: boolean;
  bestSeller: boolean;
  newArrival: boolean;
  isPublished: boolean;
  seoTitle?: string | null;
  seoDescription?: string | null;
  specifications?: string | null;
  brandId?: number | null;
  categoryId?: number | null;
  brand?: Brand | null;
  category?: Category | null;
  images: ProductImage[];
  createdAt: string;
  updatedAt: string;
}